Fix `apply` on elisp functions with no required args
When calling these functions (by saying their name directly, or via
execute) we assume functions with no required arguments and only a
rest argument take 2 args. This makes things like + and concat work
naturally for a Forth programmer. However, it should still be possible
to use `apply` to pass a custom number of args to them.